Do not load moviepy dependency directly
Loading `moviepy` too early makes the use of zou as a dependency painful. It's
because `moviepy` requires FFMPEG binary and proper configuration to access it.
When you run simple commands outside the development environment, there is no
need for FFMPEG. So no need to ask the user to configure it!