Fixing formality / grammar / punctuation of documentation
All the changes were applied to the README.md file. Made documentation more formal by replacing instances of "we", "your", etc. The author section is left the way it is to allow for better expressiveness from the authors. Fixed a few grammar and punctuation mistakes. Made the "features" list lower case, as each point should complete the sentence.