[tasks] Fix selection scrolling glitch
The selection auto scrolling was broken due to the fact that the task
panel disappeared for a millisecond causing the scroll to change before
reapparing. This led to an annoying scrolling behaviour. The tab
disappeared because to make selection cleaner, it starts by removing all
the selection before setting the new ones. It made the task panel
disappeared.
With this commit, I made a hack that keeps the tab open during the
deselection /selection process.